Implement_add_hookspec_dict in add_hookspecs
Currently, register calls _register_dict if namespace is a dict. It seems like this functionality was also intended for add_hookspecs, but _add_hookspec_dict is never called anywhere in either napari or napari_plugin_engine. The four functions' git blame patterns lead me to believe this was an accident.